+++ HISTORY.md
@@ -1,3 +1,6 @@
 # node-pg-rev - Revision History
 
+- 2022-07-25: v0.1.1 - Initial release
+  - Added defaultFetchDate to processBatch and transformRecords (ensures date before running query is available to transform)
+
 - 2022-07-24: v0.1.0 - Initial release